+++ tests/lyParser.ts
@@ -4,23 +4,8 @@ import path from "path";
 
 //import * as lilyParser from "../inc/lilyParser";
 import loadLilyParser from "../backend/loadLilyParserNode";
+import walkDir from "../backend/walkDir";
 
-
-
-const walkDir = (dir, pattern) => {
-	const list = fs.readdirSync(dir);
-	//console.log("files:", files);
-
-	const files = list.map(filename => {
-		const file = path.resolve(dir, filename);
-		const stat = fs.statSync(file);
-
-		return {file, filename, stat};
-	}).filter(({_, filename, stat}) => !stat.isDirectory() && (!pattern || pattern.test(filename)))
-		.map(({file}) => file);
-
-	return files;
-};
 
 
 const parse = (parser, sourceFile) => {